feat: add reader state persistence, improved chapter editing error handling, and dynamic page titles

+++ backend/src/main/java/com/manga/library/dto/ChapterDto.java
@@ -7,7 +7,7 @@ import lombok.Data;
 public class ChapterDto {
   private UUID id;
   private UUID seriesId;
-  private Integer chapterNumber;
+  private Double chapterNumber;
   private String title;
   private String coverImageUrl;
 }
